[NUI] Separate text constants for better readability
